Prevent X-Amzn-TraceId from Being Added More Than Once
Under some circumstances, the chain of calls through
`DelegatingHandler` instances can reach a particular handler more than
once. When that handler is `HttpClientXRayTracingHandler`, it produces
an X-Amzn-Trace-Id header that cannot be parsed because it is
_multiple_ headers. (As represented in a string by separating them
with ", ".)

This patch checks whether X-Amzn-Trace-Id has already been added, and
skips adding another if it already has been.