Hostname support (#38)
* Allow for hostnames

* Add error checking for dns resolve

* Log when falling back to default host

* Fix misspelling of 'invalid' in log messages

* Prevent multiple threads to update a HostEndPoints cache concurrently.

* Short circuit isIPCacheValid to not wait when write lock is held.

If a write lock on the cache is held we can deduce that the cache is
dirty. This allows the critical path to remain speedy.

* Rename Either.cs to EndPoint.cs

The Either class had been abandoned and been replaced with the more
domain specific EndPoint class without renaming the file.

* Change to use correct lock in HostEndPoint::updateCache().

Incorrectly acquired read lock in update, where a write lock is
necessary.

* Simplify DaemonConfig interface to avoid extra call to GetIPEndpoint().

The users of a DaemonConfig instance is often more interested in what
IP to send data to. Having to call GetIPEndpoint() each time was
therefore unergonomic.

Solved this by exposing an additional property on DaemonConfig that
returns a correct IP from the internal EndPoint object.

The Endpoint property remains.

* Clarified comments in IPEndPointExtension

* Adds documentation to new public methods and classes.

* Capitalise methods in HostEndPoint

* Add copyright notice to HostEndPoint and EndPoint.

* Capitalise methods of EndPoint.

* Remove unnecessary backing field in UdpSegmentEmitter.

* Add test for new parsing of hostname.

* Eliminate double cache update in HostEndPoint.

Rework of HostEndPoint so that multiple threads should not be able to
update the cache right after each other.

* Add unit tests for HostEndPoint

* Hide EndPoint constructor from public interface.

* Hide endpoint logic inside daemon config.

Simplify daemon config such that it internally parses ip address or
host names. UDPEndpoint and TCPEndpoint is settable again.

* Remove IsInitialised property from HostEndpoint

Property were never read, and therefore useless.

* Removed extra line.

* Corrected logging for when cache gets updated.

* Added warning log message when supplying HostEndpoint with empty string.

Test revealed that an empty string resolves to the address of the local
host when using Dns.GetHostAddresses().
See:
https://docs.microsoft.com/en-us/dotnet/api/system.net.dns.gethostaddresses?view=netframework-4.7.2

This might be non obvious behaviour when integrated with other system,
so the HostEndPoint will log a warning when initialised with an empty
string.

* Clarifies log message regarding daemon hostname isn't an IP.

* Adds tests for EndPoint functionality.

* Adds copyright notice.

Files addded to:
EndPointTests
HostEndPointTests

* Ensures that ServiceConnector refreshes ServiceURL.

ServiceConnector used to create ServiceURL upon instance initialization.
The url could become obsolete if the underlying EndPoint was a
HostEndPoint, because the ip might change as DNS records are updated.

This commit introduces `RefreshEndPoint()` that is called before
`_xrayClient` is used, ensuring that it always uses a correct
ServiceURL.

* Move TryParse(string, out HostEndPoint) to IPEndPointExtension

* Make EndPoint in DaemonConfig be either for host or ip.

Create a class EndPoint that contains either a IPEndPoint or HostEndPoint.
This class exposes a GetIPEndPoint that delegates the call to HostEndPoint
if it contains such or if it contains an IPEndPoint just returns it.

Also propagates the interface changes in the rest of the code base.

Still need to guarantee exactly what form the hosts are allowed in, but
example.com:2000 should work in addition to the earlier forms.

* Restructure endpoint parsing in order to clean up logging.

* Ensure UdpSegmentEmitter uses up to date IP.

UdpSegmentEmitter did not keep a reference to a DaemonConfig instance
but only an IP. This meant that it did not receive IP updates, even
if the DaemonConfig had some.

Fixed by storing a DaemonConfig instance and always resolving the IP
from it.

* Adds tests for resolution of hostnames when constructing daemonConfig.

* Fix decoration in copyright notice for IPEndPointExtension

* Speed up HostEndPoint tests

In `TestMultiThreadedUpdateSpacingWithDifferentTtl` removed testing for 2 and 8 secs.

Test ran for to long.